+++ examples/pong/pong.rs
@@ -7,9 +7,8 @@ use amethyst::{
     ecs::prelude::World,
     prelude::*,
     renderer::{
-        Camera, FilterMethod, MaterialTextureSet, PngFormat, Projection, Sprite, SpriteRender,
-        SpriteSheet, SpriteSheetHandle, Texture, TextureCoordinates, TextureMetadata,
-        WindowMessages,
+        Camera, MaterialTextureSet, PngFormat, Projection, Sprite, SpriteRender, SpriteSheet,
+        SpriteSheetHandle, Texture, TextureCoordinates, TextureMetadata, WindowMessages,
     },
     ui::{Anchor, TtfFormat, UiText, UiTransform},
 };
@@ -53,7 +52,7 @@ fn load_sprite_sheet(world: &mut World) -> SpriteSheetHandle {
         loader.load(
             "texture/pong_spritesheet.png",
             PngFormat,
-            TextureMetadata::srgb().with_filter(FilterMethod::Scale),
+            TextureMetadata::srgb_scale(),
             (),
             &texture_storage,
         )
@@ -124,9 +123,11 @@ fn initialise_camera(world: &mut World) {
             ARENA_WIDTH,
             ARENA_HEIGHT,
             0.0,
-        ))).with(GlobalTransform(
+        )))
+        .with(GlobalTransform(
             Matrix4::from_translation(Vector3::new(0.0, 0.0, 1.0)).into(),
-        )).build();
+        ))
+        .build();
 }
 
 /// Hide the cursor, so it's invisible while playing.
@@ -172,7 +173,8 @@ fn initialise_paddles(world: &mut World, sprite_sheet_handle: SpriteSheetHandle)
             width: PADDLE_WIDTH,
             height: PADDLE_HEIGHT,
             velocity: PADDLE_VELOCITY,
-        }).with(left_transform)
+        })
+        .with(left_transform)
         .build();
 
     // Create right plank entity.
@@ -184,7 +186,8 @@ fn initialise_paddles(world: &mut World, sprite_sheet_handle: SpriteSheetHandle)
             width: PADDLE_WIDTH,
             height: PADDLE_HEIGHT,
             velocity: PADDLE_VELOCITY,
-        }).with(right_transform)
+        })
+        .with(right_transform)
         .build();
 }
 
@@ -210,7 +213,8 @@ fn initialise_ball(world: &mut World, sprite_sheet_handle: SpriteSheetHandle) {
         .with(Ball {
             radius: BALL_RADIUS,
             velocity: [BALL_VELOCITY_X, BALL_VELOCITY_Y],
-        }).with(local_transform)
+        })
+        .with(local_transform)
         .build();
 }
 
@@ -252,7 +256,8 @@ fn initialise_score(world: &mut World) {
             "0".to_string(),
             [1.0, 1.0, 1.0, 1.0],
             50.,
-        )).build();
+        ))
+        .build();
     let p2_score = world
         .create_entity()
         .with(p2_transform)
@@ -261,6 +266,7 @@ fn initialise_score(world: &mut World) {
             "0".to_string(),
             [1.0, 1.0, 1.0, 1.0],
             50.,
-        )).build();
+        ))
+        .build();
     world.add_resource(ScoreText { p1_score, p2_score });
 }